Refactor Image and IMU Publishers
The previous method of publishing IMU messages led to a possible lag in the updating of the shared time accessed by the Image Publishers. This leads to some images getting sent with duplicate timestamps, which can break some algorithms.

This issue was solved by splitting the IMU reading and time updating tasks into separate threads, and by using a condition variable to signal when a new sample has been read.

This in effect allows the image timestamps to update quickly, without risk of delay due to IMU readings, which would originally lead to duplicate image timestamps occuring if the delay is too great.